api/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==4.1.2
django-cors-headers==3.13.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==21.3
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytz==2022.4
PyYAML==6.0
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
zope.interface==5.5.0
